remove old code, refactoring, remove duplicates
