+++ taack-ui/src/main/groovy/taack/ui/dump/RawCsvTableDump.groovy
@@ -5,32 +5,12 @@ import taack.ast.type.FieldInfo
 import taack.ast.type.GetMethodReturn
 import taack.ui.base.common.Style
 import taack.ui.base.table.UiTableVisitorImpl
-import taack.ui.style.EnumStyle
-
-import java.text.DecimalFormat
-import java.text.NumberFormat
-import java.text.SimpleDateFormat
 
 @CompileStatic
 final class RawCsvTableDump extends UiTableVisitorImpl {
     final private ByteArrayOutputStream out
 
     final static char sep = '|'
-
-    @Override
-    void visitSortableFieldHeader(String i18n, String controller, String action, Map<String, ?> params, Map<String, ?> additionalParams) {
-        out << "\"${i18n?.replace(sep, (char) ':')}\"${sep}"
-    }
-
-    @Override
-    void visitRowField(EnumStyle value, Style style, String controller = null, String action = null, Long id = null) {
-        out << (value?.name ?: "") + sep
-    }
-
-    @Override
-    void visitRowField(BigDecimal value, NumberFormat numberFormat, Style style, String controller = null, String action = null, Long id = null) {
-        out << (value?.toString() ?: "") + sep
-    }
 
     @Override
     void visitSortableFieldHeader(String i18n, FieldInfo[] fields) {
@@ -57,34 +37,14 @@ final class RawCsvTableDump extends UiTableVisitorImpl {
         out << "\"${i18n.replace(sep, (char) ':')}\"${sep}"
     }
 
-    private visitRowFieldCommon(final Class type, final Object value, final String format = null, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        switch (type) {
-            case Long:
-            case Integer:
-                visitRowField((Long) value, style, controller, action, id)
-                break
-            case Double:
-            case Float:
-            case BigDecimal:
-                visitRowField((BigDecimal) value, format, style, controller, action, id)
-                break
-            case Date:
-                visitRowField((Date) value, format, style, controller, action, id)
-                break
-            default:
-                visitRowField((String) value, style, controller, action, id)
-
-        }
+    @Override
+    void visitRowField(final FieldInfo fieldInfo, final String format, final Style style) {
+        visitRowField(fieldInfo.value?.toString(), null)
     }
 
     @Override
-    void visitRowField(final FieldInfo fieldInfo, final String format = null, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        visitRowFieldCommon(fieldInfo.fieldConstraint.field.type, fieldInfo.value, format, style, controller, action, id)
-    }
-
-    @Override
-    void visitRowField(final GetMethodReturn fieldInfo, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        visitRowFieldCommon(fieldInfo.getMethod().returnType, fieldInfo.value, null, style, controller, action, id)
+    void visitRowField(final GetMethodReturn fieldInfo, final String format, final Style style) {
+        visitRowField(fieldInfo.value?.toString(), null)
     }
 
     private static String surroundCell(final String cell) {
@@ -92,24 +52,9 @@ final class RawCsvTableDump extends UiTableVisitorImpl {
     }
 
     @Override
-    void visitRowField(final String value, final Style style, final String controller = null, final String action = null, final Long id = null) {
+    void visitRowField(final String value, final Style style) {
         out << surroundCell(value)
     }
 
-    @Override
-    void visitRowField(final Long value, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        out << surroundCell(value?.toString())
-    }
 
-    @Override
-    void visitRowField(final BigDecimal value, final String format = null, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        DecimalFormat df = new DecimalFormat(format ?: "#,###.00")
-        out << surroundCell(df.format(value))
-    }
-
-    @Override
-    void visitRowField(final Date value, final String format = null, final Style style, final String controller = null, final String action = null, final Long id = null) {
-        SimpleDateFormat sdf = new SimpleDateFormat(format ?: "yyyy-MM-dd")
-        out << surroundCell(sdf.format(value))
-    }
 }
